[NuGet] Fix package resolution with Google Play Services packages.
Fixed bug #40178 - Cannot add certain google play packages to Android
template
https://bugzilla.xamarin.com/show_bug.cgi?id=40178

The Xamarin.Android.Support.v7.AppCompat NuGet package depends on
a single version of the Xamarin.Android.Support.v4 NuGet package.
When a Xamarin Google Play Services NuGet package is installed
that depends on a higher version of the Xamarin.Android.Support.v4
NuGet package the install would fail to find a version of the
AppCompat NuGet package that is compatible. An error similar to the
following would be displayed in the Package Console window:

Could not add Xamarin.GooglePlayServices.Ads.
Updating 'Xamarin.Android.Support.v4 23.1.1.0' to
'Xamarin.Android.Support.v4 23.1.1.1' failed. Unable to find a version
of 'Xamarin.Android.Support.v7.AppCompat' that is compatible with
'Xamarin.Android.Support.v4 23.1.1.1'.

The problem was that the AppCompat NuGet package is not involved
in the initial NuGet package resolution using the remote package
source so it is initially not considered for installation into the
packages directory. When a package reference is then added to the
packages.config file a search for a compatible AppCompat package was
only using the local solution's packages directory which fails causing
the install to fail.

To fix this the local solution packages directory is used first
when looking for compatible AppCompat NuGet packages and will
fallback to using the configured remote package sources if no package
is found in this directory. This is handled by creating an aggregrate
repository, which is used as a fallback, for the
MonoDevelopProjectManager. To handle a package being added to the
packages.config at this point after the original NuGet packages have
already been downloaded to the solution's packages directory needs the
MonoDevelopPackageManager to install the NuGet package if it is not
already in the local solution packages directory.